fix : STAR 컬럼 수정으로 타입 변경

+++ src/main/java/com/example/springboot/vo/ChatroomGroupVO.java
@@ -39,5 +39,5 @@ public class ChatroomGroupVO extends CommonColumn {
 
     @Comment("별점")
     @Column(name = "STAR", nullable = true)
-    private Integer star;
+    private Double star;
 }
